Read exercises from the database instead of files
When scanning the rows of the exercises table all columns in the row
must be mapped to the fields of the Exercise struct. There is no way to
simply ignore some fields using the vanilla sql driver.

Placeholder variables have been added to get around this - ignoreInt and
ignoreString - each chosen to be used for the int and string data type
respectively. These variables are used when scanning the row to fill in
the requisite number of arguments (the runtime error was saying it must
take five arguments) and are simply ignored afterwards.