Move model test into ui_test package
It should have been in here from the beginning, but got missed. A side
effect of this change is the model.cursor needs to be exported so it can
be accessible to the tests. This isn't a great way of testing the model
since it exposes internal data to the outside world, but I can improve
this in the future. Right now, it's fine.