Inject database connection at very top of app
The in-memory database used by the tests was being destroyed before the
tests had finished running. The reason for this, was the connection was
being closed and reopened at different places in the call stack. This
wasn't noticeable when the app was running and the databse file was
being used, but it became aparent when the tests started to make use of
the in-memmory database.