Update selected exercise in View instead of Update
When updating the selected exercise in the Update function, the newly
selected exercise from the list would not be rendered. Instead, the
previous exercise would be rendered.

This is because the cursor position hadn't been updated to the new value
in the Update function, but was updated when the View function was
called.

This makes sense from a conceptual perspective, since updating the
selected exercise was about rendering it, so this belongs in the View
function even if there wasn't a bug with the selection logic.